backend/requirements.txt
flask
flask-cors
duckduckgo-search
newspaper3k
lxml_html_clean
scikit-learn
joblib
numpy
textblob
gunicorn
